Roster: roster vector now decimal[] KP-2934
Former-commit-id: 9b4700b3003f4963293d221a48937fa1d45dd564

+++ src/Core/BoundedContexts/CAPI/WB.Core.BoundedContexts.CAPI/Views/InterviewDetails/SelectebleQuestionViewModel.cs
@@ -46,7 +46,7 @@ namespace WB.Core.BoundedContexts.Capi.Views.InterviewDetails
             }
         }
 
-        public override IQuestionnaireItemViewModel Clone(int[] propagationVector)
+        public override IQuestionnaireItemViewModel Clone(decimal[] propagationVector)
         {
             IList<AnswerViewModel> newAnswers = new List<AnswerViewModel>();
             foreach (AnswerViewModel answerViewModel in this.Answers)
