+++ src/Core/SharedKernels/DataCollection/WB.Core.SharedKernels.DataCollection/Events/Interview/MultipleOptionsLinkedQuestionAnswered.cs
@@ -9,9 +9,9 @@ namespace WB.Core.SharedKernels.DataCollection.Events.Interview
 {
     public class MultipleOptionsLinkedQuestionAnswered : QuestionAnswered
     {
-        public int[][] SelectedPropagationVectors { get; private set; }
+        public decimal[][] SelectedPropagationVectors { get; private set; }
 
-        public MultipleOptionsLinkedQuestionAnswered(Guid userId, Guid questionId, int[] propagationVector, DateTime answerTime, int[][] selectedPropagationVectors)
+        public MultipleOptionsLinkedQuestionAnswered(Guid userId, Guid questionId, decimal[] propagationVector, DateTime answerTime, decimal[][] selectedPropagationVectors)
             : base(userId, questionId, propagationVector, answerTime)
         {
             this.SelectedPropagationVectors = selectedPropagationVectors;
